requirements.txt
atproto>=0.0.55
python-decouple>=3.8
python-dateutil>=2.8.2
flask>=3.0.0
gunicorn>=21.2.0
